+++ CHANGELOG.md
@@ -39,6 +39,7 @@
 #### Experts
 - `intelmq.bots.experts.sieve.expert`:
   - For `:contains`, `=~` and `!~`, convert the value to string before matching avoiding an exception. If the value is a dict, convert the value to JSON (PR#2500 by Sebastian Wagner).
+  - Add support for variables in Sieve scripts (PR#2514 by Mikk Margus Möll, fixes #2486).
 - `intelmq.bots.experts.filter.expert`:
   - Treat value `false` for parameter `filter_regex` as false (PR#2499 by Sebastian Wagner).